# docs/api/openapi.yaml
components:
  schemas:
...
    Body_upload_backup_api_v1_backup_upload_post:
      properties:
        file:
          format: binary
          title: File
          type: string
      required:
      - file
      title: Body_upload_backup_api_v1_backup_upload_post
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: Machine learning system for detecting anomalies in children's drawings
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Children's Drawing Anomaly Detection System
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/backup/download/{backup_name}:
    get:
      description: Download a specific backup file.
      operationId: download_backup_api_v1_backup_download__backup_name__get
      parameters:
      - in: path
        name: backup_name
        required: true
        schema:
          title: Backup Name
          type: string
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Download backup file
      tags:
      - backup
  /api/v1/backup/full:
    post:
      description: Create a full system backup including database and files.
      operationId: create_full_backup_api_v1_backup_full_post
      parameters:
      - description: Include uploaded files and generated content
        in: query
        name: include_files
        required: false
        schema:
          default: true
          description: Include uploaded files and generated content
          title: Include Files
          type: boolean
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create full system backup
      tags:
      - backup
...
  /api/v1/backup/restore:
    post:
      description: Restore system from a backup file.
      operationId: restore_from_backup_api_v1_backup_restore_post
      parameters:
      - in: query
        name: backup_name
        required: true
        schema:
          title: Backup Name
          type: string
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Restore from backup
      tags:
      - backup
  /api/v1/backup/upload:
    post:
      description: Upload a backup file for restoration.
      operationId: upload_backup_api_v1_backup_upload_post
      requestBody:
        content:
          multipart/form-data:
            example: {}
            schema:
              $ref: '#/components/schemas/Body_upload_backup_api_v1_backup_upload_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Upload backup file
      tags:
      - backup
  /api/v1/backup/{backup_name}:
    delete:
      description: Delete a specific backup file.
      operationId: delete_backup_api_v1_backup__backup_name__delete
      parameters:
      - in: path
        name: backup_name
        required: true
        schema:
          title: Backup Name
          type: string
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete backup file
      tags:
      - backup
...
  /api/v1/export:
    post:
      description: Export system data in specified format.
      operationId: export_data_api_v1_export_post
      parameters:
      - description: Export format
        in: query
        name: format
        required: false
        schema:
          default: json
          description: Export format
          pattern: ^(json|csv)$
          title: Format
          type: string
      - description: Include embedding vectors
        in: query
        name: include_embeddings
        required: false
        schema:
          default: false
          description: Include embedding vectors
          title: Include Embeddings
          type: boolean
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Export system data
      tags:
      - backup
  /api/v1/export/download/{export_name}:
    get:
      description: Download an exported data file.
      operationId: download_export_api_v1_export_download__export_name__get
      parameters:
      - in: path
        name: export_name
        required: true
        schema:
          title: Export Name
          type: string
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Download exported data
      tags:
      - backup
...
servers:
- description: Development server
  url: http://localhost:8000
- description: API v1 base path
  url: /api/v1